+++ packages/integration-sdk-core/src/types/jobState.ts
@@ -4,6 +4,11 @@ export interface GraphObjectFilter {
   _type: string;
 }
 
+export interface GraphObjectLookupKey {
+  _type: string;
+  _key: string;
+}
+
 export type GraphObjectIteratee<T> = (obj: T) => void | Promise<void>;
 
 /**