+++ packages/integration-sdk-runtime/src/execution/__tests__/jobState.test.ts
@@ -8,17 +8,18 @@ import {
 import { v4 as uuid } from 'uuid';
 import { FileSystemGraphObjectStore } from '../../storage';
 import { vol } from 'memfs';
-import { Entity } from '@jupiterone/integration-sdk-core';
+import {
+  Entity,
+  IntegrationMissingKeyError,
+} from '@jupiterone/integration-sdk-core';
 
 jest.mock('fs');
 
 function getMockCreateStepJobStateParams(): CreateStepJobStateParams {
-  const duplicateKeyTracker = new DuplicateKeyTracker();
-
   return {
     stepId: uuid(),
-    graphObjectStore: new FileSystemGraphObjectStore({ duplicateKeyTracker }),
-    duplicateKeyTracker,
+    graphObjectStore: new FileSystemGraphObjectStore(),
+    duplicateKeyTracker: new DuplicateKeyTracker(),
     typeTracker: new TypeTracker(),
     dataStore: new MemoryDataStore(),
   };
@@ -60,5 +61,39 @@ describe('#createStepJobState', () => {
 
     const result = await jobState.addEntities(entities);
     expect(result).toBe(entities);
+  });
+});
+
+describe('#getEntity', () => {
+  afterEach(() => {
+    vol.reset();
+  });
+
+  test('should get entity by _key', async () => {
+    const params = getMockCreateStepJobStateParams();
+    const jobState = createStepJobState(params);
+    const entity: Entity = {
+      _type: 'a_entity',
+      _class: 'A',
+      _key: 'a',
+    };
+
+    await jobState.addEntity(entity);
+    expect(await jobState.getEntity('a')).toStrictEqual(entity);
+  });
+
+  test('should throw if entity does not exist', async () => {
+    expect.assertions(2);
+    const params = getMockCreateStepJobStateParams();
+    const jobState = createStepJobState(params);
+
+    try {
+      await jobState.getEntity('invalid-entity-key');
+    } catch (err) {
+      expect(err instanceof IntegrationMissingKeyError).toEqual(true);
+      expect(err.message).toEqual(
+        `Failed to find entity in in-memory graph object metadata store (_key=invalid-entity-key)`,
+      );
+    }
   });
 });